fix new transaction
